[misc] align with master
* up to "Core: Fix #122 warning: no previous prototype for ‘usbi_log_v’"
  a8e2e7cab278fbfcba91909eaf762cf66da9eda5

+++ libusb/descriptor.c
@@ -188,7 +188,7 @@ static int parse_interface(libusb_context *ctx,
 	int len;
 	int r;
 	int parsed = 0;
-	int tmp;
+	size_t tmp;
 	struct usb_descriptor_header header;
 	struct libusb_interface_descriptor *ifp;
 	unsigned char *begin;
@@ -336,7 +336,7 @@ static int parse_configuration(struct libusb_context *ctx,
 	int i;
 	int r;
 	int size;
-	int tmp;
+	size_t tmp;
 	struct usb_descriptor_header header;
 	struct libusb_interface *usb_interface;
